Speed up interp loop when running on behalf of JIT.
Now about 4x faster but still at least 2x slower than raw interp.
Shouldn't be a huge difference as JIT doesn't call interp in most hot paths.

+++ interp.h
@@ -9,7 +9,7 @@ struct interp_struct;
 
 struct cpu_driver* interp_create(struct cpu_driver_funcs* p_funcs);
 
-int interp_enter_with_details(
+int64_t interp_enter_with_details(
     struct interp_struct* p_interp,
     int64_t countdown,
     int (*instruction_callback)(void* p,
